refactor(stella-cli): split the agent_engine_config cluster out of settings.rs
settings.rs crossed the 1500-line ratchet at 1542 when the steering switch
landed (#3348's red file-size check). The engine-agent cluster — the largest
self-contained concept in the file — moves verbatim to settings/engine.rs,
re-exported so every settings:: path keeps resolving. The one visibility
change: the three overlay methods become pub(super), which sibling merge.rs
was already reaching as a parent-module private.

Refs #3348
